+++ src/symbols/enum_based.rs
@@ -1,4 +1,3 @@
-use crate::float_ops::BasicNumOps;
 use crate::ops::*;
 use crate::*;
